library-dev/project_1/tests/model/factory/test_condition_evaluator.py
import pytest
import pandas as pd
from unittest.mock import patch, MagicMock

from src.model.factory.condition_evaluator import ConditionEvaluator
from src.lib.validator_utils.ibr_decision_table_validator import DT
from src.lib.common_utils.ibr_enums import LogLevel

from src.model.factory.editor_factory import EditorFactory
from unittest.mock import Mock, MagicMock
from src.model.facade.base_facade import DataFrameEditor

from src.lib.common_utils.ibr_dataframe_helper import tabulate_dataframe

# config共有
import sys
from src.lib.common_utils.ibr_decorator_config import initialize_config
config = initialize_config(sys.modules[__name__])
log_msg = config.log_message
from src.lib.common_utils.ibr_enums import LogLevel

class TestConditionEvaluatorInit:
    """ConditionEvaluatorの__init__メソッドのテスト

    テスト構造:
    ├── C0: 基本機能テスト
    │   └── dt_functionsの初期化確認
    ├── C1: 分岐カバレッジ
    │   └── DT クラスに適切なメソッドがある場合とない場合
    ├── C2: 条件網羅
    │   └── DT クラスの様々なメソッド組み合わせ
    └── BVT: 境界値テスト
        └── DT クラスにメソッドがない極端なケース

    C1のディシジョンテーブル:
    | 条件                           | ケース1 | ケース2 |
    |--------------------------------|---------|---------|
    | DT クラスにメソッドが存在する  | Y       | N       |
    | 出力                           | 正常初期化| 空の辞書 |

    境界値検証ケース一覧：
    | ケースID | 入力パラメータ | テスト値               | 期待される結果 | テストの目的/検証ポイント           | 実装状況 | 対応するテストケース              |
    |----------|-----------------|-----------------------|----------------|--------------------------------------|----------|-----------------------------------|
    | BVT_001  | DT クラス       | メソッドなし          | 空の辞書       | DTクラスにメソッドがない場合の処理  | 実装済み | test_init_BVT_no_methods          |
    | BVT_002  | DT クラス       | 1つのメソッドのみ     | 1要素の辞書    | 最小限のメソッド数での初期化        | 実装済み | test_init_BVT_single_method       |
    | BVT_003  | DT クラス       | 多数のメソッド(100個) | 100要素の辞書  | 大量のメソッドがある場合の処理      | 実装済み | test_init_BVT_many_methods        |

    境界値検証ケースの実装状況サマリー：
    - 実装済み: 3
    - 未実装: 0
    - 一部実装: 0

    注記：
    すべての境界値検証ケースが実装されています。これらのテストにより、__init__メソッドの動作が様々な条件下で正しく機能することを確認しています。
    """

    @pytest.fixture
    def mock_config(self):
        config = MagicMock()
        config.log_message = MagicMock()
        return config

    def test_init_C0_basic_functionality(self, mock_config):
        test_doc = """
        テスト内容:
        - テストカテゴリ: C0
        - テスト区分: 正常系
        - テストシナリオ: dt_functionsの初期化確認
        """
        log_msg(f"\n{test_doc}", LogLevel.INFO)

        evaluator = ConditionEvaluator()

        assert hasattr(evaluator, 'dt_functions')
        assert isinstance(evaluator.dt_functions, dict)
        assert all(callable(func) for func in evaluator.dt_functions.values())
